backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.19.0
httpx>=0.27.2
postgrest==2.19.0
realtime==2.0.2
storage3==0.7.5
supabase-auth==2.19.0
supabase-functions==2.19.0
python-dotenv==1.0.1
pydantic==2.9.2
requests>=2.31.0
openai